Python 2.x Unicode handling fix
The code could fail if a unicode object is used as the query, and a
utf-8 encoded string with characters outside of the ASCII range is
passed in as one of the query arguments. This fix encodes the query in
the appropriate character set *before* applying the escaped args, as the
MySQLdb adapter does.