Make Connect show through help from inner class
The redirection seems a faff when wanting to quickly look things up while working with pymysql. Grabbing the doco from the inner class (which should always be correct given it is a passthrough) should give a nice link. 

I've removed the "_orig_conn" from the local namespace (del never deletes anything, it just removes that name/ref for it). This keeps it from fouling up the dir(..) output.